Refactor: Use uint8_t for motor speed percentage parameter
This refines the `speedPercent` parameter type for motor `deploy` and `retract` methods from `int` to `uint8_t`. This change improves type safety and ensures that speed percentages are correctly constrained to a non-negative range (0-255), aligning with their intended use as percentage values.

+++ src/driver/motors.cpp
@@ -1,6 +1,7 @@
 #include "motor.h"
 #include "motors.h"
 
+// Public
 Motors::Motors(Motor m1, Motor m2)
   : motor1(m1), motor2(m2) {}
 
@@ -9,12 +10,12 @@ void Motors::init() {
   motor2.init();
 }
 
-void Motors::deploy(int speedPercent) {
+void Motors::deploy(uint8_t speedPercent) {
   motor1.deploy(speedPercent);
   motor2.deploy(speedPercent);
 }
 
-void Motors::retract(int speedPercent) {
+void Motors::retract(uint8_t speedPercent) {
   motor1.retract(speedPercent);
   motor2.retract(speedPercent);
 }
